Remove regression_tests/css4_color_test/generate_css4_module.py
